[scripts/show_xmas_tree.sh] fix to kill a job when only the job is running

+++ scripts/show_xmas_tree.sh
@@ -5,4 +5,8 @@
 htop
 
 jobs_num=`jobs | sed -n -e 's/^\[\([0-9]*\)\]+\s*Running\s*\/vagrant\/xmas_tree_workload\s.*$/\1/p'`
-kill %${jobs_num}
+if [ "x${jobs_num}" != x ]; then
+    kill %${jobs_num}
+else
+    exit 1
+fi
